Create s.sh
Upon compilation of the mta.c file, you will find an a.out file. That is the a.out referred to in this file. This shell script takes in a MIDI file as input and gives the file output.txt which contains the notes along with the duration for which it is pressed.